+++ src/ImageModalTrigger.php
@@ -401,12 +401,13 @@ class ImageModalTrigger {
 	 */
 	private function getWidthOptionForThumbLimits( $thumbLimits ) {
 
+		// this could also be affected by issue #9
 		$user = RequestContext::getMain()->getUser();
 		$widthOption = $user::getDefaultOption( 'thumbsize' );
 
 		// we have a problem here: the original \Linker::makeImageLink does get a value for $widthOption,
 		// for instance in parser tests. unfortunately, this value is not passed through the hook.
-		// so there are instances, there $thumbLimits[$widthOption] is not defined.
+		// so there are instances, where $thumbLimits[$widthOption] is not defined.
 		// solution: we cheat and take the first one
 		if ( $widthOption !== null && isset( $thumbLimits[$widthOption] ) ) {
 			return $widthOption;